Added BaseViewModel. Wired with DashboardViewModel.
Currently all view models share the same navigation logic.
